Initial revision of Spending Screen
