+++ templates/Spending.html
@@ -1,0 +1,24 @@
+{% extends "Base.html" %}
+{% block head %}
+<style>
+	
+</style>
+{% endblock %}
+{% block tabs %}
+<li id="loading_tab" class="tab">
+	<a><h6>Loading...</h6></a>
+</li>
+{% endblock %}
+
+{% block content %}
+
+{% endblock %}
+
+{% block script %}
+<script type="module">
+	import SpendingController from './static/js/controller/spendingController.js'
+	document.addEventListener("DOMContentLoaded", () => {
+		new SpendingController().init();
+	});
+</script>
+{% endblock %}
